pyproject.toml
[tool.poetry]
name = "agentic-data-analyst"
version = "0.1.0"
description = ""
authors = ["imnotdev25 <85677268+imnotdev25@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.11.10"
fastapi = "^0.115.5"
pydantic-settings = "^2.6.1"
crewai = "^0.80.0"
crewai-tools = "^0.14.0"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
